+++ android/src/main/kotlin/com/example/emulator_checker/EmulatorCheckerPlugin.kt
@@ -1,0 +1,58 @@
+package com.example.emulator_checker
+
+import android.os.Build
+import androidx.annotation.NonNull
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.plugin.common.MethodCall
+import io.flutter.plugin.common.MethodChannel
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler
+import io.flutter.plugin.common.MethodChannel.Result
+import android.util.Log
+
+/** EmulatorCheckerPlugin */
+class EmulatorCheckerPlugin: FlutterPlugin, MethodCallHandler {
+  private lateinit var channel : MethodChannel
+
+  override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "emulator_checker")
+    channel.setMethodCallHandler(this)
+  }
+
+  override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
+    when (call.method) {
+      "isEmulator" -> result.success(isEmulator())
+      else -> result.notImplemented()
+    }
+  }
+
+  private fun isEmulator(): Boolean {
+    val radioVersion = getRadioVersion()
+    Log.d("EmulatorChecker", "Radio Version: $radioVersion")
+
+    return (Build.FINGERPRINT.startsWith("generic") ||
+            Build.FINGERPRINT.lowercase().contains("unknown") ||
+            Build.MODEL.contains("Emulator") ||
+            Build.MODEL.contains("Android SDK built for x86") ||
+            Build.MANUFACTURER.contains("Genymotion") ||
+            Build.BRAND.startsWith("generic") && Build.DEVICE.startsWith("generic") ||
+            "google_sdk" == Build.PRODUCT ||
+            Build.HARDWARE.contains("goldfish") ||
+            Build.HARDWARE.contains("ranchu") ||
+            Build.PRODUCT.contains("sdk") ||
+            Build.PRODUCT.contains("vbox86p") ||
+            Build.PRODUCT.contains("emulator") ||
+            radioVersion == null || radioVersion.isEmpty() || radioVersion == "1.0.0.0")
+  }
+
+  private fun getRadioVersion(): String? {
+    return try {
+      Build.getRadioVersion()
+    } catch (e: Exception) {
+      null
+    }
+  }
+
+  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+    channel.setMethodCallHandler(null)
+  }
+}