+++ example/emulator_checker_example.dart
@@ -1,7 +1,5 @@
 import 'package:emulator_checker/emulator_checker.dart';
 
 void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-  print('EmulatorChecker: ${EmulatorChecker().isEmulator()}');
+  print('EmulatorChecker: ${EmulatorChecker.isEmulator()}');
 }