+++ example/ios/Runner/GeneratedPluginRegistrant.m
@@ -1,0 +1,21 @@
+//
+//  Generated file. Do not edit.
+//
+
+// clang-format off
+
+#import "GeneratedPluginRegistrant.h"
+
+#if __has_include(<emulator_checker/EmulatorCheckerPlugin.h>)
+#import <emulator_checker/EmulatorCheckerPlugin.h>
+#else
+@import emulator_checker;
+#endif
+
+@implementation GeneratedPluginRegistrant
+
++ (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
+  [EmulatorCheckerPlugin registerWithRegistrar:[registry registrarForPlugin:@"EmulatorCheckerPlugin"]];
+}
+
+@end