+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.0.2
+- 1.0.2 add android platform specific code for radio version checker
+
+## 1.0.1
+- 1.0.1 emulator checker platform specific code
+
+## 1.0.0
+- 1.0.0 update to null safety
+
+## 0.0.4
+- 0.0.4 updated readme to cater null safety.
+
 ## 0.0.3
 - 0.0.2 updated the device_info_plus to latest, reverted back to 4.0.0